feat: re-use impl for HashmapAug && PfxHashmap
