+++ crates/tlb-ton/src/lib.rs
@@ -1,12 +1,12 @@
 mod address;
 mod boc;
 mod currency;
-mod hashmap;
+pub mod hashmap;
 mod message;
 mod state_init;
 mod timestamp;
 mod unary;
 
 pub use self::{
-    address::*, boc::*, currency::*, hashmap::*, message::*, state_init::*, timestamp::*, unary::*,
+    address::*, boc::*, currency::*, message::*, state_init::*, timestamp::*, unary::*,
 };
